+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ### Additions
 
 - Added a generic hash map for generic keys and values (`misc/genhashmap.h`).
+- Added a settings struct.
+- Added debug rendering for colliders.
 
 ### Removals